requirements.txt
pandas
openpyxl
google-generativeai
google-api-python-client
requests